Refactored newbeard.py.
