+++ utils/newbeard.py
@@ -4,6 +4,8 @@ import argparse
 from pathlib import Path
 from shutil import copytree
 
+from .utils import make_init_text
+
 import stringcase
 
 
@@ -16,32 +18,8 @@ def make_readme(dir_, name):
 def make_init(dir_, name):
     python_path = dir_ / Path("python/{}".format(name))
     python_path.mkdir(parents=True)
-    init_text = '''
-from skybeard.beards import BeardChatHandler
-from skybeard.decorators import onerror
-from skybeard.utils import get_args
 
-
-class {beardclassname}(BeardChatHandler):
-
-    __userhelp__ = """Default help message."""
-
-    __commands__ = [
-        # command, callback coro, help text
-        ("echo", 'echo', 'Echos arguments. e.g. <code>/echo [arg1 [arg2 ... ]]</code>')
-    ]
-
-    # __init__ is implicit
-
-    @onerror
-    async def echo(self, msg):
-        args = get_args(msg)
-        if args:
-            await self.sender.sendMessage("Args: {{}}".format(args))
-        else:
-            await self.sender.sendMessage("No arguments given.")
-
-    '''.strip().format(beardclassname=stringcase.pascalcase(name))
+    init_text = make_init_text(name)
 
     with (python_path / Path("__init__.py")).open("w") as f:
         f.write(init_text)
